Basic implementation of `/feed` endpoint.
This serves an OPDS feed of a user's Marked for Later list
at the `/feed` endpoint. It can be called by a logged-in user,
or by an anonymous client if they
pass the user's AO3 credentials as GET parameters.
(These are not stored by the app, although the server software
may log these parameters. It is not very safe!)

Next steps: Providing each feed with a secret sharable address.