Supporting bookmark, subscription, history feeds.
This commit refactors the code from `marked_for_later` to be
more generic, allowing each new
feed to be implemented simply by wrapping one function call.

This also expands the entries in `FEED_TYPES`, which means that
all of the new feed types are automatically presented to the user
via the management interface and on the top-level index.